Update startslide scripts and inputs
