+++ scripts/MKW_Inputs/Startslides/mach_left.csv
@@ -1,0 +1,401 @@
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,1,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,1,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+0,0,0,0,0,1,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,1,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+0,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,1,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-2,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,-3,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,7,0,1,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
