+++ scripts/MKW_Inputs/Startslides/wiggle_right.csv
@@ -1,0 +1,407 @@
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,1,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,1,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+0,0,0,7,0,0,-1
+0,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,1,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,7,0,0,-1
+1,0,0,-2,0,0,-1
+1,0,0,4,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,1,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,-7,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
+1,0,0,0,0,0,-1
